Add redraw HOC
