+++ src/index.js
@@ -4,6 +4,7 @@ export * from './icons'
 export { default as theme, themeDark, brand, colors } from './theme'
 export { font, grid, spring, breakpoint, BreakPoint, unselectable } from './utils/styles'
 export { formatIntegerRange } from './utils/format'
+export { default as redraw } from './components/HOC/redraw'
 export { default as BaseStyles } from './components/BaseStyles/BaseStyles'
 export { default as Section } from './components/Section/Section'
 export { default as IllustratedSection } from './components/IllustratedSection/IllustratedSection'
